pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docent"
version = "0.1.0"
description = "Provider-agnostic natural language AI interface for scenario-driven modelling"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40.0"]
azure = ["openai>=1.0.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
docent = "docent.adapters.cli.main:cli"
docent-mcp = "docent.adapters.mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/docent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
